requirements.txt
# Core dependencies with flexible versions
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0,<2.0.0
torch>=2.0.0
sentence-transformers>=2.2.0
transformers>=4.30.0
faiss-cpu>=1.7.0
hdbscan>=0.8.0
scikit-learn>=1.3.0
rapidfuzz>=3.0.0
plotly>=5.0.0

